requirements.txt
black
isort
flake8
pylint
abstract-singleton
wheel
setuptools
build
twine
auto_gpt_plugin_template

# Testing
pytest